ScheduledRecordingService tested
